requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
execnet==1.9.0
idna==3.2
iniconfig==1.1.1
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-forked==1.3.0
pytest-xdist==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
tabulate==0.8.9
toml==0.10.2
urllib3==1.26.6
